refacto: various fixes
