+++ @vates/types/src/xo.mts
@@ -356,7 +356,7 @@ export type XoBackupJob = BaseXoJob & {
   mode: 'full' | 'delta'
   name?: string
   remotes?: {
-    id: XoRemote['id'] | { __or: XoRemote['id'][] }
+    id: XoBackupRepository['id'] | { __or: XoBackupRepository['id'][] }
   }
   vms?: {
     id: XoVm['id'] | { __or: XoVm['id'][] } | Record<string, unknown>
@@ -391,10 +391,6 @@ export type XoBackupJob = BaseXoJob & {
   }
 }
 export type XoJob = BaseXoJob & {}
-
-export type XoRemote = {
-  id: Branded<'remote'>
-}
 
 export type XoSchedule = {
   cron: string
@@ -596,7 +592,7 @@ export type XapiXoRecord =
   | XoVmTemplate
   | XoVtpm
 
-export type NonXapiXoRecord = XoGroup | XoProxy | XoJob | XoBackupRepository | XoSchedule | XoServer | XoUser | XoRemote
+export type NonXapiXoRecord = XoGroup | XoProxy | XoJob | XoBackupRepository | XoSchedule | XoServer | XoUser
 
 export type XoRecord = XapiXoRecord | NonXapiXoRecord
